# Add WrapErr to ErrorEvent