--- a/error_event.go
+++ b/error_event.go
@@ -36,16 +36,27 @@ var errorEventPool = &sync.Pool{
 // Please see the example at LTSVLogger.Err for an example usage.
 type ErrorEvent struct {
 	error
-	buf []byte
+	originalErr error
+	buf         []byte
 }
 
 // Err creates an ErrorEvent with the specified error.
 func Err(err error) *ErrorEvent {
 	e := errorEventPool.Get().(*ErrorEvent)
 	e.error = err
+	e.originalErr = err
 	e.buf = e.buf[:0]
-	e.buf = append(e.buf, "err:"...)
-	e.buf = append(e.buf, escape(err.Error())...)
+	return e
+}
+
+// WrapErr wraps an ErrorEvent or a plain error and returns a new error.
+func WrapErr(err error, wrapper func(err error) error) *ErrorEvent {
+	e, ok := err.(*ErrorEvent)
+	if !ok {
+		return Err(err)
+	}
+
+	e.error = wrapper(e.error)
 	return e
 }
 
@@ -218,14 +229,22 @@ func (e *ErrorEvent) UTCTime(label string, value time.Time) *ErrorEvent {
 	return e
 }
 
-// Error returns the error string with labeled values in the LTSV format.
+// Error returns the error string without labeled values.
 func (e *ErrorEvent) Error() string {
-	return string(e.buf)
+	return e.error.Error()
+}
+
+// ErrorWithValues returns the error string with labeled values.
+func (e *ErrorEvent) ErrorWithValues() string {
+	buf := make([]byte, 0, 8192+len(e.buf))
+	buf = append(buf, "err:"...)
+	buf = append(buf, escape(e.Error())...)
+	return string(append(buf, e.buf...))
 }
 
 // OriginalError returns the original error.
 func (e *ErrorEvent) OriginalError() error {
-	return e.error
+	return e.originalErr
 }
 
 // appendStack appends a formated stack trace of the calling goroutine to buf