--- a/log.go
+++ b/log.go
@@ -261,9 +261,10 @@ func (l *LTSVLogger) Err(err error) {
 	if !ok {
 		errorEvent = Err(err)
 	}
-	buf := make([]byte, 8192)
+	errStr := errorEvent.ErrorWithValues()
+	buf := make([]byte, 0, 64+len(errStr))
 	buf = l.appendPrefixFunc(buf[:0], "Error")
-	buf = append(buf, errorEvent.buf...)
+	buf = append(buf, errStr...)
 	buf = append(buf, '\n')
 	_, _ = l.writer.Write(buf)
 	errorEventPool.Put(errorEvent)