This pull request causes a performance degradation, but it's needed for ErrorEvent
to be used without calling ltsvlog.Logger.Err.

```
name                      old time/op    new time/op    delta
Info-2                      1.64µs ± 1%    1.59µs ± 1%    -2.86%  (p=0.000 n=10+10)
ErrWithStackAndUTCTime-2    18.6µs ± 8%    21.6µs ± 2%   +16.09%  (p=0.000 n=10+10)
ErrWithStack-2              17.8µs ± 5%    21.3µs ± 3%   +19.54%  (p=0.000 n=10+10)
ErrWithUTCTime-2            5.54µs ± 4%    7.24µs ± 2%   +30.80%  (p=0.000 n=10+10)

name                      old alloc/op   new alloc/op   delta
Info-2                       0.00B          0.00B           ~     (all equal)
ErrWithStackAndUTCTime-2    8.26kB ± 0%   16.53kB ± 0%  +100.16%  (p=0.000 n=10+10)
ErrWithStack-2              8.26kB ± 0%   16.53kB ± 0%  +100.17%  (p=0.000 n=10+10)
ErrWithUTCTime-2            8.21kB ± 0%   16.46kB ± 0%  +100.57%  (p=0.000 n=10+10)

name                      old allocs/op  new allocs/op  delta
Info-2                        0.00           0.00           ~     (all equal)
ErrWithStackAndUTCTime-2      3.00 ± 0%      5.00 ± 0%   +66.67%  (p=0.000 n=10+10)
ErrWithStack-2                3.00 ± 0%      5.00 ± 0%   +66.67%  (p=0.000 n=10+10)
ErrWithUTCTime-2              2.00 ± 0%      4.00 ± 0%  +100.00%  (p=0.000 n=10+10)
```